% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfree.R
\name{complement}
\alias{complement}
\alias{reverse}
\alias{reverse_complement}
\title{Reversing and complementing degenerate oligos}
\usage{
complement(strings, codes = IUB_CODES)

reverse(strings)

reverse_complement(strings, codes = IUB_CODES)
}
\description{
Reversing and complementing degenerate oligos
}
\section{Functions}{
\itemize{
\item \code{complement}: Complement degenerate oligos

\item \code{reverse}: Reverse a degenerate oligo

\item \code{reverse_complement}: Reverse complement of a degenerate oligo
}}

