% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfree.R
\name{cutfree}
\alias{cutfree}
\title{Design degenerate oligos free of restriction sites}
\usage{
cutfree(len = 20, sites = c(), codes = IUB_CODES,
  starting_oligo = strrep("N", len), min_blocks = 1,
  obj_weights = log(1:4), re_randomize = TRUE, seed = NULL,
  obj_frac = 1, quiet = FALSE, maxtime = 30)
}
\arguments{
\item{len}{Length of the random oligo. Must be provided if
\code{starting_oligo} is not given.}

\item{sites}{Character vector of restriction sites to block.}

\item{starting_oligo}{Starting oligo from which sites will be
removed. If not given, defaults to a string of N's.}

\item{min_blocks}{Minimum number of blocks at each site, i.e.
the minimum number of changes that need to be made for a cut
site to appear anywhere in the oligo.}

\item{obj_weights}{Objective function weights for each code's
degeneracy.}

\item{re_randomize}{If \code{TRUE} (default), re-run the optimization
to randomize the codes while maintaining the same number of oligos.}

\item{seed}{Seed for the random number generator.}

\item{obj_frac}{Fraction of the objective function that must be
maintained during re-randomization.}

\item{quiet}{Run silently with no output from Gurobi.}

\item{maxtime}{Maximum time (in seconds) allowed for the solver.}
}
\value{
A list containing \code{code} -- the randomize oligo, and a
fields describing the MILP and the solution statistics.
}
\description{
Design pools of DNA oligos that are guaranteed to not
contain specified restriction enzyme cut sites. All cut
sites are blocked while ensuring the optimal number of oligos
remain in the pool.
}
\examples{
result <- cutfree(m=20, sites=c("GGTCTC", "GATATC"), min_blocks=2)
print_degenerate_oligo(result$code)

}

