% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfree.R
\name{degeneracy}
\alias{degeneracy}
\title{Calculate number of sequences in a degerate oligo}
\usage{
degeneracy(sequence, codes = IUB_CODES)
}
\description{
Calculate number of sequences in a degerate oligo
}

