% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfree.R
\docType{data}
\name{IUB_CODES}
\alias{IUB_CODES}
\title{IUB codes used by CutFree}
\format{An object of class \code{list} of length 15.}
\usage{
IUB_CODES
}
\description{
IUB codes used by CutFree
}
\keyword{datasets}

