% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutfree.R
\name{make_oligo_block}
\alias{make_oligo_block}
\alias{print_oligo_block}
\title{Text summary of degenerate oligos}
\usage{
make_oligo_block(oligo, codes = IUB_CODES)

print_oligo_block(oligo, indent = "")
}
\description{
Create textual representations of an oligo describing the base content at
each position.
}
\details{
\code{make_oligo_block} creates a text block showing which bases are allowed
by the oligo. It returns four strings denoting the presence
of each base.

\code{print_oligo_block} prints the output of
\code{make_oligo_block} plus the oligo.
}
\examples{
make_oligo_block("ACGTRYMKSWHBVDN")

print_oligo_block("ACGTRYMKSWHBVDN")

}

